ADD: Fucntional Todo's. Everything works now!
The user can create proejcts and new todos. Todos can be deleted as well as projects. deleting a projects removes all of the tasks which were associated with it. Had to redesign the architecture and make projects into another dictionary in order to avoid being flabbergasted by ids that would keep on changing.