Fixed Issue #30
To fix issue #30, of custom controls not working on Fedora 34/KDE, I simply created a new scene that is ran when ever the project is launched. When it does this, it gets the operating system and if you are running an X11 or Server system, then it disables the custom controls, and enables the window manager default controls.